feature(model): move impl class to helpers.py
